🚚 rename: BasicUserInfo 위치를 index로 변경 #47
